+++ CHANGELOG.md
@@ -48,6 +48,7 @@ All notable changes to this project will be documented in this file.
 - AS direct editation/highlighter §§pop and §§dup can have getmember or call
 - [#1295], [#116] AS3 dynamic construction via MultinameL
 - ABC cleaner not properly fixing namespace sets
+- The hex view does not display bytes if there are too few of them
 
  ### Changed
 - Icon of "Deobfuscation options" menu from pile of pills to medkit